+++ core.lua
@@ -1635,8 +1635,8 @@ end
 
 --- Used as a "DoCellUpdate" function for lib-st
 function RCLootCouncil.SetCellClassIcon(rowFrame, frame, data, cols, row, realrow, column, fShow, table, class)
-	local celldata = data[realrow].cols and data[realrow].cols[column] or data[realrow][column]
-	local class = celldata.args and celldata.args[1] or class
+	local celldata = data and (data[realrow].cols and data[realrow].cols[column] or data[realrow][column])
+	local class = celldata and celldata.args and celldata.args[1] or class
 	if class then
 		frame:SetNormalTexture("Interface\\GLUES\\CHARACTERCREATE\\UI-CHARACTERCREATE-CLASSES"); -- this is the image containing all class icons
 		local coords = CLASS_ICON_TCOORDS[class]; -- get the coordinates of the class icon we want